Rewrite TLS Caddy files on install
Porting this from master branch for now, makes it a lot easier for
people to go from master to 1.1.* since the master branch has a
different Caddyfile structure.

Since Caddy 0.9.* still has the X-Accel-Redirect issue re:
Content-Type, we can't tag on master yet, so this will help people get
back to a tagged release without having to manually fix a bunch of
Caddyfiles.