Change `valet domain` command to `valet tld`
Fixes #144
".test" or ".dev" is really a TLD, not a "domain" in the conventional sense.
Changing the command to `valet tld` more accurately reflects the purpose of the command (to set or get the configured TLD served by Valet)

The use of `valet domain` is currently preserved as an alias for `valet tld`, but will be removed at a later date.