Set files path as a static path for concrete5 driver
Previously testing a concrete5 site that had tons of 404s and required file based sessions would force each request to load sequentially causing the site to load pretty slowly. This change makes it so that the driver is aware of the files directory and markes anything within as a static file.