Modernise Unit Tests
Modernise referenes to brew commands to ones that could be run manually as a confirmation today.

mcrypt has been removed and bundled into php@7.1 and above so it has been replaced with the realistic array example where both php and php@7.2 are returned.

The exception are the tap commands which have always been completely illustrative.