Fix merge error
The domain-name determination using strrpos wasn't very efficient. Stripping the `.crt` from the end of the filename is easier to parse.
Now all it returns is the first part of the string up until the final `.`, treating the remainder as whatever tld was used for the domain, regardless of what tld valet is configured for.