Update Site.php
Update the domain() function to include the default TLD only if the $domain variable is not given explicitly. If the $domain variable is given, just return that instead of sticking the default TLD on top. This feature lets you secure custom domains, e.g. "valet secure www.mywebsite.com" (if you also add www.mywebsite.com to your hosts file and create a symlink or folder in the Valet site directory, of course).