Enable securing custom domains
Update the domain() function to skip including the default TLD if there is a dot in the domain name. This lets you secure custom domains, e.g. "valet secure www.mywebsite.com" (if you also add www.mywebsite.com to your hosts file and create a symlink or folder in the Valet site directory, of course).