Improve link command's secure and isolate sub-commands
Passing a custom hostname to the `link` command with the `--isolate` argument will result in either a mismatched site being isolated or the site not being found ("The [example] site could not be found in Valet's site list.").

The reason for this is because the `isolate` command defaults to using the current directory's _name_ which won't match the custom hostname.

This issue does not affect the `secure` command because it resolves the hostname from the current directory's _full path_.

This commit passes the hostname (either custom or the current directory's name) to both sub-commands to ensure consistency of operations.

Note: A bug introduced in mnapoli/silly v1.8.1 breaks all arguments in `runCommand()` method. See mnapoli/silly#69 for details and mnapoli/silly#70 for pull request.