Bandage for 502 errors using Gettext/PostgreSQL/MongoDB
Many have reported that these changes "do" help the problem.
And yet many others have reported that they make no difference.

Closes #1433

This PR is merely a bandage for a broader problem that exists in the PHP build for MacOS. See the issue referenced above for discussion and links to Homebrew and PHP repository discussions. The problem is not in Homebrew, but in the PHP build published by the PHP core.
Chime in on the PHP discussion if you want to stir up more active participation for a proper fix, telling them that you're a Mac user encountering the fork segfaults (they won't care whether you're using Valet, but they will recognize if you mention Homebrew).