fix(php-fpm): correctly reconfigure all versions
This previously only reapplied the configuration for the linked version.
Because it is recommended to run valet install again after a brew upgrade, it is necessary to update all utilized PHP versions.

On brew upgrades, some versions received the default www.conf again. This would leave the user with a broken setup until they either call `valet use php@X.X` for all the versions or isolate a site again with the version.

This commit fixes the behavior by rerunning the configuration step on all php versions. This is more consistent, because in the next line all utilized PHP versions are also restarted.